# wallet: Keep track of the wallet's own transaction outputs in memory