--- a/src/wallet/wallet.cpp
+++ b/src/wallet/wallet.cpp
@@ -4459,4 +4459,14 @@ void CWallet::RefreshAllTXOs()
         RefreshTXOsFromTx(wtx);
     }
 }
+
+std::optional<WalletTXO> CWallet::GetTXO(const COutPoint& outpoint) const
+{
+    AssertLockHeld(cs_wallet);
+    const auto& it = m_txos.find(outpoint);
+    if (it == m_txos.end()) {
+        return std::nullopt;
+    }
+    return it->second;
+}
 } // namespace wallet