… src/bench/wallet_create_tx.cpp
… src/wallet/test/wallet_tests.cpp
… src/wallet/wallet.h
… src/wallet/walletdb.cpp
… test/functional/wallet_balance.py