@remyers 2023-09-15
Concept ACK

My preliminary testing confirmed that this PR will significantly speed up common operations like `fundrawtransaction` on wallets with many years of transaction history. A single `fundrawtransaction` RPC call that took 5 sec on a simulated large wallet now takes 50ms. Services providers with large wallets that need to make many RPC calls per minute will benefit from reduced lock contention on `cs_wallet`; for example, LSPs using bitcoin core wallets like Eclair.

Caching the wallet's tx outputs in memory seems like a straight forward solution as long as the cache of tx outs in `m_txos` are always updated (or regenerated) if the underlying set of wallet txs in `mapWallet` or the `m_spk_managers` change.