@remyers 2023-09-15
Exposing `RefreshWalletTXOs()` as a public method puts the responsibility of knowing when it should be called outside of the `CWallet` scope. What do you think about making it private and clearing `m_txos` (or adding a new flag) to trigger `RefreshWalletTXOs()` at query time?

This alternate approach would require `GetTXO(const COutPoint& outpoint)` and `GetTXOs()` to first call `RefreshWalletTXOs()` if `m_txos` is empty.  There are only a few methods that change `m_spk_managers` that would need to clear `m_txos` to trigger a one time `RefreshWalletTXOs()`:
 - `ImportPrivKeys`
 - `ImportPublicKeys`
 - `ImportScriptPubKeys`
 - `AddScriptPubKeyMan`

It's also only a small set of places where keys are imported in `backup.cpp`, so the question of leaving `RefreshWalletTXOs()` public might just be a matter of which approach would be less foot-gunny in the future and if there are disadvantages to delaying the regeneration of txos.

I think you can safely make `RefreshWalletTxTXOs(const CWalletTx& wtx)` a private method now though because it is only called by wallet methods that add a wtx to `mapWallet` like `LoadToWallet` and `AddToWallet`.