@achow101 2023-09-15
> Exposing `RefreshWalletTXOs()` as a public method puts the responsibility of knowing when it should be called outside of the `CWallet` scope. What do you think about making it private and clearing `m_txos` (or adding a new flag) to trigger `RefreshWalletTXOs()` at query time?

I'm not sure that's meaningfully different. The caller still needs to know when to clear `m_txos` in order to trigger a future `RefreshWalletTXOs()`. The way it is now also provides a better UX as the slow part of refreshing the TXOs occurs during slow import operations rather than the first time those TXOs are needed after the import.