@remyers 2023-09-16
> I'm not sure that's meaningfully different. The caller still needs to know when to clear `m_txos` in order to trigger a future `RefreshWalletTXOs()`. The way it is now also provides a better UX as the slow part of refreshing the TXOs occurs during slow import operations rather than the first time those TXOs are needed after the import.

Just to clarify, I mean that a reviewer, but not method caller, would need to make sure every method in CWallet that invalidates `m_txos` also clears it. I agree with your point though that forcing the refresh at import time makes sense. I also notice that there are a number of places where a `ScriptPubKeyMan` gets added or updated outside of CWallet, so I understand now why it's not really possible to make `RefreshWalletTXOs()` something internal to `CWallet`.