@w0xlt 2025-04-02
This PR also simplifies the `GetBalance` and `AvailableCoins` methods. 
The introduction of the `GetTXO()` method also makes the wallet interface clearer. 
As far as I could verify, the `m_txo` in-memory  cache is reliably kept in sync.

ACK https://github.com/bitcoin/bitcoin/pull/27286/commits/48545bdb3b9f59a8b78640dbf20afff6e2663ad7