@davidgumberg 2025-06-06
reACK https://github.com/bitcoin/bitcoin/commit/ec25de35b2a51b493d36452f75096f790230a4e9

Reviewed range-diff (`git range-diff 3973425...ec25de35`), addresses all the nits I left in my previous review that were correct.  Mainly, the benchmark iterations are no longer unnecessarily hardcoded, two transaction level checks that were being performed per-output in `AvailableCoins()` are now part of the cached tx-level checks, and stylistically, code was moved to more closely resemble the order of checks in master, making the (whitespace-hidden) diff much smaller, and the changes in this PR easier to remove in my opinion: e.g. (https://github.com/bitcoin/bitcoin/commit/854241298059f8b56563ca7ca64f5b61b59f119b?w=1 vs https://github.com/bitcoin/bitcoin/pull/27286/commits/2703d84b7efafcce05387a5e018363c3c59c4b58?w=1)

I benchmarked again (w/ different hardware[^1])  and it looks like the performance gap in the  `AvailableCoins()` benchmark is improved over the earlier branch (10% faster than master vs 5% faster than master) 

<details>

<summary>

### Benchmark results

</summary>

```console
$ ./build/bin/bench_bitcoin -min-time=30000 '-filter=Wallet(Available|CreateTxUse).*'
```

#### pr branch (https://github.com/bitcoin/bitcoin/commit/ec25de35b2a51b493d36452f75096f790230a4e9)

|               ns/op |                op/s |    err% |          ins/op |          cyc/op |    IPC |         bra/op |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|       39,666,039.18 |               25.21 |    0.2% |  743,701,163.19 |  184,433,998.97 |  4.032 |  13,191,744.00 |    0.1% |     32.98 | `WalletAvailableCoins`
|          818,711.43 |            1,221.43 |    0.1% |   16,376,793.42 |    3,821,152.74 |  4.286 |     264,358.23 |    0.2% |     33.07 | `WalletCreateTxUseOnlyPresetInputs`
|       28,955,730.17 |               34.54 |    0.7% |  370,601,527.15 |  135,128,956.19 |  2.743 |  66,153,487.66 |    0.4% |     33.26 | `WalletCreateTxUsePresetInputsAndCoinSelection`

#### master (e2174378aa8a3 + a39074c52afd073abd9101dc6) 

|               ns/op |                op/s |    err% |          ins/op |          cyc/op |    IPC |         bra/op |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|       44,211,982.81 |               22.62 |    0.2% |  803,845,700.31 |  205,309,518.10 |  3.915 |  18,081,465.08 |    0.0% |     33.04 | `WalletAvailableCoins`
|          855,419.56 |            1,169.02 |    0.1% |   16,581,860.02 |    3,991,552.91 |  4.154 |     355,371.44 |    0.2% |     33.02 | `WalletCreateTxUseOnlyPresetInputs`
|       35,074,882.27 |               28.51 |    0.2% |  505,591,550.24 |  163,578,562.91 |  3.091 |  91,987,283.54 |    0.2% |     33.16 | `WalletCreateTxUsePresetInputsAndCoinSelection`

#### old pr branch (https://github.com/bitcoin/bitcoin/commit/39734259fc10b1339db2ae80c5fab12916cf0aff) with `epochIterations` removed

|               ns/op |                op/s |    err% |          ins/op |          cyc/op |    IPC |         bra/op |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|       42,297,947.25 |               23.64 |    0.0% |  774,588,933.86 |  196,770,180.57 |  3.937 |  15,813,962.88 |    0.1% |     32.95 | `WalletAvailableCoins`
|          852,312.97 |            1,173.28 |    0.2% |   16,650,566.31 |    3,978,823.89 |  4.185 |     374,875.56 |    0.2% |     33.04 | `WalletCreateTxUseOnlyPresetInputs`
|       28,841,852.53 |               34.67 |    0.4% |  373,458,753.83 |  134,591,208.06 |  2.775 |  68,667,702.77 |    0.4% |     33.26 | `WalletCreateTxUsePresetInputsAndCoinSelection`

</details>

[^1]: AMD Ryzen 9 7900X 12-Core Processor used in these benchmarks.